fix(ui:form): rename form.ts to form-control.ts to fix case-sensitive errors
