+++ packages/ui/src/components/form/FormGroup.tsx
@@ -1,4 +1,4 @@
-import type { FormGroup } from './form';
+import type { FormGroup } from './form-control';
 
 import React from 'react';
