+++ packages/ui/src/components/form/index.ts
@@ -1,4 +1,4 @@
-export * from './form';
+export * from './form-control';
 export * from './Form';
 export * from './FormItem';
 export * from './FormGroup';
